src/requirements.txt
requests
xmltodict
python-dotenv
selenium
pyvirtualdisplay
opencv-python
pytesseract
Pillow
